+++ packages/at_client/test/local_key_decryption_test.dart
@@ -22,17 +22,14 @@ void main() {
   });
   group('A group of local key decryption tests', () {
     test('test to verify decryption of local key', () async {
-      var rsaKeyPair = AtChopsUtil.generateAtEncryptionKeyPair();
+      var rsaKeyPair = RsaKeyPair.generate();
       var sharedSymmetricKey = 'REqkIcl9HPekt0T7+rZhkrBvpysaPOeC2QL1PVuWlus=';
       var encryptedSharedSymmetricKey = EncryptionUtil.encryptKey(
           sharedSymmetricKey, rsaKeyPair.atPublicKey.publicKey);
-      AtChopsKeys atChopsKeys = AtChopsKeys.create(rsaKeyPair, null);
-      var atChopsImpl = AtChopsImpl(atChopsKeys);
       when(() => mockAtClient.getCurrentAtSign()).thenReturn('@alice');
-      when(() => mockAtClient.atChops).thenAnswer((_) => atChopsImpl);
-      print('encryptedSharedSymmetricKey:$encryptedSharedSymmetricKey');
       when(() => mockAtClient.getLocalSecondary())
           .thenReturn(mockLocalSecondary);
+      stubEncryptionKeyPair(mockLocalSecondary, '@alice', rsaKeyPair);
       when(() => mockLocalSecondary.executeVerb(any<LLookupVerbBuilder>()))
           .thenAnswer((_) => Future.value('data:$encryptedSharedSymmetricKey'));
       var localKey = AtKey()